+++ src/components/CardSlider/CardSlider.css
@@ -1,0 +1,48 @@
+.slider-wrap {
+  position: relative;
+}
+
+.like {
+  position: absolute;
+  bottom: 20px;
+  right: 20px;
+}
+
+.like-count {
+  position: absolute;
+  top: 3px;
+  right: 30px;
+  font-size: 15px;
+}
+
+.slick-slider {
+  margin-bottom: 30px;
+}
+
+.slick-dots {
+  bottom: -20px;
+}
+
+.slick-slide {
+  transition: transform 0.2s ease-in;
+}
+
+.slick-slide.slick-active {
+  transition: transform 0.2s ease-in;
+  margin: 0;
+}
+
+.slick-slide:not(.slick-active) {
+  /* transform: scale(0.87); */
+  visibility: hidden;
+}
+
+.portfolio-card {
+  position: relative;
+  margin: 10px 0;
+  padding: 20px 15px;
+  height: 200px;
+  background-color: #fff;
+  border-radius: 25px;
+  box-shadow: 0 3px 6px 0 rgba(0, 0, 0, 0.16);
+}
